Renamed module to ddmr
