+++ EvaluationScripts/Evaluate_3d_weaklySupervised.py
@@ -11,13 +11,13 @@ import voxelmorph as vxm
 import neurite as ne
 from datetime import datetime
 
-import DeepDeformationMapRegistration.utils.constants as C
-from DeepDeformationMapRegistration.data_generator import DataGeneratorManager
-from DeepDeformationMapRegistration.utils.misc import try_mkdir
-from DeepDeformationMapRegistration.utils.nifti_utils import save_nifti
-from DeepDeformationMapRegistration.networks import WeaklySupervised
-from DeepDeformationMapRegistration.losses import HausdorffDistanceErosion
-from DeepDeformationMapRegistration.layers import UncertaintyWeighting
+import ddmr.utils.constants as C
+from ddmr.data_generator import DataGeneratorManager
+from ddmr.utils.misc import try_mkdir
+from ddmr.utils.nifti_utils import save_nifti
+from ddmr.networks import WeaklySupervised
+from ddmr.losses import HausdorffDistanceErosion
+from ddmr.layers import UncertaintyWeighting
 
 
 os.environ['CUDA_DEVICE_ORDER'] = C.DEV_ORDER
