+++ ddmr/layers/uncertainty_weighting.py
@@ -12,9 +12,9 @@ import tensorflow as tf
 import numpy as np
 import random
 
-from DeepDeformationMapRegistration.utils.operators import soft_threshold, gaussian_kernel, sample_unique
-import DeepDeformationMapRegistration.utils.constants as C
-from DeepDeformationMapRegistration.utils.thin_plate_splines import ThinPlateSplines
+from ddmr.utils.operators import soft_threshold, gaussian_kernel, sample_unique
+import ddmr.utils.constants as C
+from ddmr.utils.thin_plate_splines import ThinPlateSplines
 from voxelmorph.tf.layers import SpatialTransformer
 from neurite.tf.utils import resize
 #from cupyx.scipy.ndimage import zoom
