+++ ddmr/utils/operators.py
