+++ demo/requirements.txt
@@ -1,2 +1,2 @@
-deepdeformationmapregistration @ git+https://github.com/jpdefrutos/DDMR.git
+ddmr @ git+https://github.com/jpdefrutos/DDMR.git
 gradio==3.44.4
